moved mixer ionto its own folder
