+++ src/main/java/frc/robot/subsystems/mixer/Mixer.java
@@ -1,7 +1,8 @@
-package frc.robot.subsystems;
+package frc.robot.subsystems.mixer;
 
 import com.ctre.phoenix.motorcontrol.can.WPI_TalonSRX;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
+import frc.robot.subsystems.MoveableSubsystem;
 
 import static frc.robot.Robot.robotConstants;
